+++ src/components/SmartParkingSolution/ArrowTextSubComponent/index.tsx
